Add files via upload
A  “sentiment”  is  a   generally  binary  opposition  in  opinions  and  expresses the feelings   in the  form  of  emotions,  attitudes,  opinions,  and so on. It can express many opinions.Natural  Language  Processing  (NLP)  is  an area  which make machines , capable to understand natural languages  e.g. English , Chinese  etc. Sentiment  Analysis  is subfield of NLP, it studies that how machines process text and   perform   classification   of  text . 
Vader stands for Valence Aware Dictionary and sEntiment Reasoner. It is a lexicon and rule based tool for sentiment analysis. It is specifically attuned to sentiments expressed in social media.It is used for analyzing the sentiment of text which contains both positive and negative polarity. The main function of VADER is to quantify how much of positive or negative emotion is present in the text. It can also measure the intensity of emotion.VADER is an unsupervised learning algorithm widely used in Sentiment Analysis.
RoBERTa stands for “Robustly Optimized BERT pre-training Approach”. In many ways this is a better version of the BERT model.RoBERTa is a transformers model pretrained on a large corpus of English data in a self-supervised fashion. This means it was pretrained on the raw texts only, with no humans labelling them in any way (which is why it can use lots of publicly available data) with an automatic process to generate inputs and labels from those texts. https://huggingface.co/arpanghoshal/EmoRoBERTa.